% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adresseToCoord.R
\name{adresseToCoord}
\alias{adresseToCoord}
\title{Geolocaliser des adresses et renvoyer leurs coordonnees en WGS84 (EPSG 4326)}
\usage{
adresseToCoord(
  adresses,
  nbEchos = 1,
  codePostal = NULL,
  codeInsee = NULL,
  interactive = FALSE
)
}
\arguments{
\item{adresses}{vecteur texte Adresses à géolocaliser.}

\item{nbEchos}{valeur numérique. Nombre de résultats maximum par adresse en cas de doute sur la géolocalisation. Par défaut 1.}

\item{codePostal}{vecteur texte. Par défaut NULL.}

\item{codeInsee}{vecteur texte. Par défaut NULL.}

\item{interactive}{booléen. Choix du contexte d'exécution. Si TRUE, contexte shiny. Par défaut FALSE.}
}
\value{
Un data.frame
}
\description{
La fonction adresseToCoord permet de récupérer les coordonnées WGS84 (longitude/latitude code epsg 4326) de plusieurs adresses.

La fonction utilise l'API Adresses (documentation : https://geo.api.gouv.fr/adresse).
}
\details{
En plus des coordonnées lon et lat, un score entre 0 et 1 est proposé, indiquant la pertinence de la géolocalisation.

Le codePostal et le codeInsee sont des arguments optionnels qui permettent d'améliorer la géolocalisation en cas de doute sur un libellé d'adresse.

Le code postal peut aussi être mentionné directement dans les libellés d'adresses.

Le nombre d'echos permet à la fonction de retourner jusqu'à n echos si la géolocalisation renvoie plusieurs possibilités. Seuls les échos les plus pertinents sont proposés.
}
\examples{
# Exemple 1 : avec une seule adresse
adresseToCoord(adresses = "88 avenue Verdier Montrouge",
               nbEchos = 1)

# Exemple 2 : avec un vecteur d'adresses
adresses <- c("1 Rue des Abeilles 13001 Marseille",
              "1 Allee des Abeilles 13016 Marseille",
              "1 Impasse Abeille 13003 Marseille",
              "1 Impasse de la Chapelle 13013 Marseille",
              "1 Boulevard de la Chapelle 13009 Marseille",
              "1 Boulevard de la Chapelle 13014 Marseille")

adresseToCoord(adresses = adresses,
               nbEchos = 1) # un resultat par adresse

adresseToCoord(adresses = adresses,
               nbEchos = 2) # 2 resultats max possibles par adresse

}
