% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeComToCoord.R
\name{codeComToCoord}
\alias{codeComToCoord}
\title{Fonction de passage entre code commune INSEE et coordonnees au chef-lieu (chx) ou au centroide ou a un point sur la commune (pos).}
\usage{
codeComToCoord(
  codeInsee,
  geo = format(Sys.time(), format = "\%Y"),
  type = "chx"
)
}
\arguments{
\item{codeInsee}{vecteur texte.}

\item{geo}{texte. Par défaut, l'année courante.}

\item{type}{texte. Type de point souhaité. A choisir parmi "chx" par défaut (chef-lieu de la commune), "centroide" (barycentre) ou "pos" (point sur la surface).}
}
\value{
Un data.frame de trois colonnes "code", "lon" et "lat".
}
\description{
La fonction codeComToCoord permet de récupérer des coordonnées WGS84 (EPSG 4326) de plusieurs communes à partir des codes communes INSEE.

Les points en sortie correspondent aux chefs-lieux de communes (chx), aux centroïdes ou à des points situés obligatoirement à l'intérieur de la commune,
calculés selon l'algorithme spécifique "point_on_surface".

Selon la morphologie du contour de la commune, le centroïde peut être situé en-dehors de ses limites.
Pour que le point soit obligatoirement situé dans la commune, il faut alors spécifier type = "pos" (point_on_surface).

Il est possible de choisir une géographie de 2017 à l'année de mise à jour du package.
Par exemple, si le package a été mis à jour en géographie 2023, il peut transformer en coordonnées lon/lat des codes communes des COG 2017 à 2023.
}
\examples{
# Renvoie les coordonnees des chefs-lieux des communes de Montrouge (92049)
# et de Malakoff (92046) en geographie 2021.
codeComToCoord(codeInsee = c("92049","92046"),
               geo = "2020",
               type = "chx")

# Renvoie les coordonnees des centroides des communes de Montrouge (92049)
# et de Malakoff (92046) en geographie 2021.
codeComToCoord(codeInsee = c("92049","92046"),
               geo = "2020",
               type = "centroide")

# Renvoie les coordonnees "point_on_surface" des communes de Montrouge (92049)
# et de Malakoff (92046) en geographie 2021.
codeComToCoord(codeInsee = c("92049","92046"),
               geo = "2020",
               type = "pos")

# Renvoie les coordonnees des chefs-lieux des communes d'Ancteville (50007)
# et de Saint-Sauveur-Lendelin (50550) en geographie 2018.
codeComToCoord(codeInsee = c("50007","50550"),
               geo = "2018",
               type = "chx")

# Au 01/01/2019, les communes d'Ancteville (50007),
# de Saint-Sauveur-Lendelin (50550) et autres communes
# ont fusionne pour former Saint-Sauveur-Villages (50550).

# Renvoie les coordonnees du chef-lieu de la commune
# de Saint-Sauveur-Villages (50550) en geographie 2019.
# Et affiche un avertissement si le code commune est introuvable
# pour la geographie donnee.
codeComToCoord(codeInsee = c("50007","50550"),
               geo = "2019",
               type = "chx")

}
