% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertTo.R
\name{convertTo}
\alias{convertTo}
\title{Convertir des coordonnees d'un systeme geographique a un autre et dans un format different (sp, sf ou data.frame)}
\usage{
convertTo(from, to = NULL, fromEpsg = NULL, toEpsg = 4326, interactive = FALSE)
}
\arguments{
\item{from}{objet sf ou sp avec un CRS valide, ou un data.frame de deux colonnes (coordonnées) ou 3 colonnes (id en 1ère colonne puis les coordonnées).}

\item{to}{texte. Type d'objet souhaité en sortie : "sf", "sp" ou "data.frame". Si NULL (par défaut), le type de sortie sera le même que celui de l'objet en entrée.}

\item{fromEpsg}{texte ou numérique. Si from est un data.frame, fromEpsg doit correspondre au code EPSG des coordonnées à convertir.}

\item{toEpsg}{texte ou numérique. Code EPSG souhaité pour l'objet à convertir.}

\item{interactive}{booléen. Choix du contexte d'exécution. Si TRUE, contexte shiny. Par défaut FALSE.}
}
\value{
Un objet sf, sp ou un data.frame
}
\description{
La fonction convertTo permet de convertir un objet sf, sp ou un data.frame de coordonnées dans un système géographique différent et dans un autre format, au choix entre sp, sf ou data.frame.
}
\details{
Le code EPSG du système WGS84 est communément 4326.

En France métropolitaine, le système de projection en vigueur est le Lambert 93, code EPSG 2154.

Pour la Guadeloupe et la Martinique, le code EPSG est 5490 pour la projection UTM 20 N.

Pour la Guyane, le code EPSG est 2972 pour la projection UTM 22 N.

Pour la Réunion, le code EPSG est 2975 pour la projection UTM 40 S.

Pour Mayotte, le code EPSG est 4471 pour la projection UTM 38 S.

La projection Mercator pour représenter la mappemonde a pour code EPSG 3395.
}
\examples{
# Creation d'un data.frame de 4 coordonnees en Lambert 93.
coord <- data.frame(id = c(1:4),
                    X = c(897740.5,901367.8,874261.9,897740.5),
                    Y = c(6272912,6251706,6291801,6272912),
                    stringsAsFactors = FALSE)

# Conversion des coordonnees en WGS84 (EPSG 4326).
# Transformation du data.frame en objet spatial sf.
coord_sf_WGS84 <- convertTo(from = coord,
                            to = "sf",
                            fromEpsg = 2154,
                            toEpsg = 4326)

# Transformation de l'objet sf en objet spatial sp.
coord_sp_WGS84 <- convertTo(from = coord_sf_WGS84,
                            to = "sp")

# Transformation de l'objet sp en data.frame.
# Conversion des coordonnees en Lambert 93 (EPSG 2154).
coord_dt_L93 <- convertTo(from = coord_sp_WGS84,
                          to = "data.frame",
                          toEpsg = 2154)

# Creation d'un objet sf : un point en coordonnees Lambert93 (EPSG 2154).
objet_sf_points <- sf::st_sf(geometry = sf::st_sfc(
                                        sf::st_geometry(
                                        sf::st_point(
                                        c(897740.5,6272912.0)
                                        )),
                             crs=2154))

# Transformation de l'objet sf en data.frame.
# Conversion implicite des coordonnees en WGS84 (EPSG 4326).
coord_dt_WGS84 <- convertTo(from = objet_sf_points,
                            to = "data.frame")

}
