% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indIsoSpatial.R
\name{indIsoSpatial}
\alias{indIsoSpatial}
\title{Calculer des indicateurs en volume par regroupement de mailles selon leur appartenance spatiale}
\usage{
indIsoSpatial(
  pol,
  maille,
  var,
  pond = NULL,
  choixIntersect = 1,
  return = "pol",
  idGroup = NULL
)
}
\arguments{
\item{pol}{objet sf (POLYGON ou MULTIPOLYGON) ou sp (SpatialPolygons ou SpatialPolygonsDataFrame).}

\item{maille}{objet sf (POLYGON ou MULTIPOLYGON) ou sp (SpatialPolygonsDataFrame).}

\item{var}{vecteur texte. Noms des variables en volume à sommer de l'objet maille.}

\item{pond}{vecteur numérique. Noms des variables de pondération de l'objet maille. A appliquer aux variables à sommer. Par défaut NULL.}

\item{choixIntersect}{numérique. 1, 2 ou 3 (voir details). Par défaut 1.}

\item{return}{vecteur texte. "pol", "maille" ou c("pol","maille") (voir details). Par défaut "pol".}

\item{idGroup}{texte. Nom de la variable identifiant les pol et constituant les groupes de maille si return contient "maille". Par défaut NULL.}
}
\value{
Si return == "pol", l'objet pol, enrichi des variables sommées issues de l'objet maille.

Si return == "maille", l'objet maille, enrichi d'une variable indiquant le groupe d'appartenance de la maille à pol et les valeurs sommées pour chaque groupe.

Si return == c("pol","maille"), une liste contenant pol et maille est renvoyée, enrichie des variables décrites ci-dessus.
}
\description{
La fonction indIsoSpatial permet de sommer une ou plusieurs variables (la population par exemple)
d'une maille (maille communale ou grille par exemple) selon leur appartenance spatiale à une zone particulière (polygones d'isochrones par exemple).
}
\details{
Si les variables de pondérations sont renseignées, le vecteur des variables de pondérations doit respecter le même ordre que le vecteur des variables à sommer.

Le paramètre choixIntersect vaut 1, 2 ou 3. Si une maille intersecte pol :

choix 1 : la totalité de la variable de la maille intersectée est comptabilisée dans la somme.

choix 2 : la maille est tronquée selon le découpage de pol. Les valeurs à sommer sont recalculées selon le prorata de la surface de la maille tronquée.

choix 3 : la totalité de la variable de la maille intersectée est exclue de la somme.

Les arguments pol et maille doivent être dans le même système de projection, peu importe le système.
}
\examples{
# Specification d'un serveur osrm obligatoire pour executer les exemples
options(osrm.server = "https://metric-osrm-backend.lab.sspcloud.fr/")

# Specification du profil
options(osrm.profile = "driving")

# Appel a la fonction metricOsrmIso
iso <- metricOsrmIso(loc = data.frame(lon = 4.92,lat = 46.15))

# Conversion en projection metrique Lambert 93 (EPSG 2154)
iso_L93 <- convertTo(from = iso[[1]],
                     toEpsg = 2154)

# Construction d'une grille
grille <- sf::st_as_sf(x = sf::st_make_grid(iso_L93, cellsize = 52000))

# Ajout d'une variable en volume a la grille
grille$vol <- c(15,11,26,5,3,17)

# Calcul de l'indicateur spatial
ind_res <- indIsoSpatial(pol = iso_L93,
                         maille = grille,
                         var = "vol",
                         choixIntersect = 1,
                         return = c("pol","maille"),
                         idGroup = "id")

}
