% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indTableSrcDst.R
\name{indTableSrcDst}
\alias{indTableSrcDst}
\title{Calculer des indicateurs en volume par source (src) ou par destination (dst)}
\usage{
indTableSrcDst(res)
}
\arguments{
\item{res}{data.frame. Résultat de la fonction metricOsrmTable avec nbDstVolOiseau = 1 ou nbDstMeasure = 1.}
}
\value{
liste de deux objets sf.

Le 1er élément comporte une source (src) par observation. Les variables en volume sont couplées à une variable indiquant la part de ce volume dans l'ensemble des sources ayant la même destination (en pourcentage).

Le 2ème élément comporte une destination (dst) par observation (voir details).
}
\description{
La fonction indTableSrcDst permet de calculer par destination (dst) la somme d'une ou plusieurs variables en volume issues des sources (src), la population par exemple.

La fonction permet également de calculer la part des variables en volume pour chaque source dans l'ensemble des sources ayant une même destination.
}
\details{
La table res doit etre calculée à partir de la fonction metricOsrmTable avec les arguments nbDstVolOiseau = 1 ou nbDstMeasure = 1.
En effet, il ne faut qu'un seul résultat par source : la destination la plus proche d'une source à vol d'oiseau, en temps ou en distance.

L'ordre des colonnes de res en entrée de indTableSrcDst doit être obligatoirement le même que celui obtenu en sortie de la fonction metricOsrmTable.

Pour calculer les indicateurs, il faut ajouter des variables en volume à la suite du tableau de résultats (la population par source par exemple).

En sortie, le 2ème élément de la liste indique les indicateurs en volume par destination. Il s'agit de la somme de toutes les variables de sources pour chaque destination.
Il donne également un comptage du nombre de sources allant vers chaque destination.
Si il n'y a pas de variable en volume ajoutée au tableau de résultats, seul l'indicateur de comptage est indiqué.

Les deux éléments sont des objets sf. Il est alors possible, par exemple, de réaliser une carte en ronds proportionnels par destination selon un indicateur calculé.
}
\examples{
# Reproduction d'une table de resultats avec la fonction metricOsrmTable
res <- data.frame(ID = c(1:4),
                  idSrc = c("1","2","3","4"),
                  lonSrc = c(5.39200,5.39242,5.37107,5.46476),
                  latSrc = c(43.28292,43.28368,43.47900,43.31246),
                  idDst = c("B","B","A","C"),
                  lonDst = c(5.38385,5.38385,5.47678,5.38219),
                  latDst = c(43.28571,43.28571,43.29028,43.44144),
                  duree = c(87.3,62.8,726.8,239.7),
                  distance = c(1103.3,942.7,8318.6,3252.6),
                  pop = c(1204,806,1164,976),
                  stringsAsFactors = FALSE)

res_ind <- indTableSrcDst(res = res)

}
