% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.osrm-package.R
\docType{package}
\name{metric.osrm-package}
\alias{metric.osrm}
\alias{metric.osrm-package}
\title{metric.osrm: Distancier Metric OSRM}
\description{
Calculs de temps et distances depuis un serveur OSRM. Temps d'acces, accessibilite et isochrones.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://git.lab.sspcloud.fr/metric-osrm/metric-osrm-package/-/issues}
}

}
\author{
\strong{Maintainer}:  PSAR-AT Unité CED - INSEE PACA \email{dr13-psar-at@insee.fr}

Other contributors:
\itemize{
  \item  SED - DR Hauts-de-France - INSEE [contributor]
  \item  SED - DR Normandie - INSEE [contributor]
  \item  SED - DR Bourgogne-Franche-Comté - INSEE [contributor]
  \item  Institut national de la statistique et des études économiques [copyright holder]
  \item Timothée Giraud (\href{https://orcid.org/0000-0002-1932-3323}{ORCID}) (Multiple code snippets, from osrm package) [copyright holder]
}

}
\keyword{internal}
