% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statTable.R
\name{statTable}
\alias{statTable}
\title{Calculer des statistiques sur les resultats de la fonction metricOsrmTable.}
\usage{
statTable(res)
}
\arguments{
\item{res}{data.frame. Résultat de la fonction metricOsrmTable avec les colonnes duree et/ou distance.}
}
\value{
Une liste de valeurs statistiques
}
\description{
La fonction statTable permet de calculer des statistiques simples sur les résultats de la fonction metricOsrmTable (comptages, valeurs min, valeurs max, moyennes, médianes).
}
\details{
La table res doit correspondre à la table de résultats de la fonction metricOsrmTable avec les colonnes suivantes :
"ID","idSrc","lonSrc","latSrc","idDst","lonDst","latDst", "duree" et/ou "distance".

Les statistiques calculées sont :

- le nombre de couples calculés ;

- le nombre de sources (src) ;

- le nombre de destinations (dst).

Pour la duree et la distance :

- la valeur min ;

- la valeur max ;

- la moyenne ;

- la médiane.
}
\examples{
# Reproduction d'une table de resultats avec la fonction metricOsrmTable
res <- data.frame(ID = c(1:4),
                  idSrc = c("1","2","3","4"),
                  lonSrc = c(5.39200,5.39242,5.37107,5.46476),
                  latSrc = c(43.28292,43.28368,43.47900,43.31246),
                  idDst = c("B","B","A","C"),
                  lonDst = c(5.38385,5.38385,5.47678,5.38219),
                  latDst = c(43.28571,43.28571,43.29028,43.44144),
                  duree = c(87.3,62.8,726.8,239.7),
                  distance = c(1103.3,942.7,8318.6,3252.6),
                  pop = c(1204,806,1164,976),
                  stringsAsFactors = FALSE)

res_ind <- statTable(res = res)

}
